Fix font size/position ugliness

+++ include/Nooskewl_Engine/font.h
@@ -27,9 +27,9 @@ public:
 	void enable_shadow(SDL_Colour shadow_colour, Shadow_Type shadow_type);
 	void disable_shadow();
 
-	void draw(SDL_Colour colour, std::string text, Point<int> dest_position);
+	void draw(SDL_Colour colour, std::string text, Point<float> dest_position);
 	// Returns number of characters drawn, plus whether or not it filled the max in bool &full
-	int draw_wrapped(SDL_Colour colour, std::string text, Point<int> dest_position, int w, int line_height, int max_lines, int started_time, int delay, bool dry_run, bool &full, int &num_lines, int &width);
+	int draw_wrapped(SDL_Colour colour, std::string text, Point<float> dest_position, int w, int line_height, int max_lines, int started_time, int delay, bool dry_run, bool &full, int &num_lines, int &width);
 
 private:
 	void cache_glyph(Uint32 ch);
